(BSR)[PRO] refactor: merge collective factories into one file
