+++ pro/src/pages/CollectiveOffers/__specs__/CollectiveOffersQueryParams.spec.tsx
@@ -9,10 +9,10 @@ import { CollectiveOfferResponseModel, OfferStatus } from 'apiClient/v1'
 import { DEFAULT_SEARCH_FILTERS } from 'core/Offers/constants'
 import { SearchFiltersParams } from 'core/Offers/types'
 import { computeCollectiveOffersUrl } from 'core/Offers/utils'
+import { collectiveOfferFactory } from 'utils/collectiveApiFactories'
 import { renderWithProviders } from 'utils/renderWithProviders'
 
 import { CollectiveOffers } from '../CollectiveOffers'
-import { collectiveOfferFactory } from '../utils/collectiveOffersFactories'
 
 //FIX ME : extract inital values and constant to reduce code duplication with CollectiveOffers.spec.tsx
