+++ pro/src/utils/collectiveApiFactories.ts
@@ -23,6 +23,8 @@ import {
   VenueTypeCode,
   GetCollectiveOfferResponseModel,
   GetCollectiveOfferTemplateResponseModel,
+  CollectiveOfferResponseModel,
+  ListOffersVenueResponseModel,
 } from 'apiClient/v1'
 import { BOOKING_STATUS } from 'core/Bookings/constants'
 
@@ -33,6 +35,37 @@ let offererId = 1
 let bookingId = 1
 let bookingDetailsId = 1
 let institutionId = 1
+
+export const collectiveOfferFactory = (
+  customCollectiveOffer: Partial<CollectiveOfferResponseModel> = {}
+): CollectiveOfferResponseModel => {
+  const currentId = offerId++
+
+  return {
+    id: currentId,
+    status: OfferStatus.ACTIVE,
+    isActive: true,
+    hasBookingLimitDatetimesPassed: true,
+    isEducational: true,
+    name: `offer name ${offerId}`,
+    venue: listOffersVenueFactory(),
+    stocks: [],
+    isEditable: true,
+    isPublicApi: false,
+    interventionArea: [],
+    isShowcase: false,
+    ...customCollectiveOffer,
+  }
+}
+export const listOffersVenueFactory = (
+  customListOffersVenue: Partial<ListOffersVenueResponseModel> = {}
+): ListOffersVenueResponseModel => ({
+  id: 1,
+  name: 'venue name',
+  offererName: 'offerer name',
+  isVirtual: false,
+  ...customListOffersVenue,
+})
 
 const sharedCollectiveOfferData = {
   isActive: true,
